feat(): moved logic into a service
